+++ playground/control/connectors/docker/client.py
@@ -1,5 +1,4 @@
 
-import base64
 import os
 from collections import OrderedDict
 from typing import Union
@@ -8,27 +7,15 @@ import aiodocker
 from aiodocker.containers import DockerContainers
 from aiodocker.networks import DockerNetworks
 
-from playground.control.attribs import (
-    ProxyCreateCommandAttribs, ProxyDeleteAttribs,
-    ServiceCreateCommandAttribs, ServiceDeleteAttribs)
-from playground.control.command import PlaygroundCommand
 from playground.control.connectors.docker.events import (
     PlaygroundDockerEvents)
 from playground.control.connectors.docker.networks import (
     PlaygroundDockerNetworks)
+from playground.control.connectors.docker.proxies import (
+    PlaygroundDockerProxies)
+from playground.control.connectors.docker.services import (
+    PlaygroundDockerServices)
 from playground.control.decorators import cmd, method_decorator
-
-
-class PlaygroundDockerServices(object):
-
-    def __init__(self, connector):
-        self.connector = connector
-
-
-class PlaygroundDockerProxies(object):
-
-    def __init__(self, connector):
-        self.connector = connector
 
 
 class PlaygroundDockerClient(object):
@@ -45,17 +32,13 @@ class PlaygroundDockerClient(object):
     @method_decorator(cmd)
     async def clear(self) -> list:
         await self.networks.clear()
-        resources = (
-            (self.list_services, self.service_delete),
-            (self.list_proxies, self.proxy_delete))
-        for _resources, remove in resources:
-            for resource in await _resources():
-                await remove(dict(name=resource['name']))
+        await self.proxies.clear()
+        await self.services.clear()
 
     @method_decorator(cmd(sync=True))
     async def dump_resources(self) -> list:
         proxies = OrderedDict()
-        for proxy in await self.list_proxies():
+        for proxy in await self.proxies.list():
             proxies[proxy["name"]] = proxy
 
         networks = OrderedDict()
@@ -63,7 +46,7 @@ class PlaygroundDockerClient(object):
             networks[network["name"]] = network
 
         services = OrderedDict()
-        for service in await self.list_services():
+        for service in await self.services.list():
             services[service["name"]] = service
         return dict(
             networks=networks,
@@ -85,147 +68,10 @@ class PlaygroundDockerClient(object):
                 return True
         return False
 
-    async def list_proxies(self) -> list:
-        return await self._list_resources(
-            self.docker.containers, "proxy")
-
-    async def list_services(self) -> list:
-        return await self._list_resources(
-            self.docker.containers, "service")
-
-    @method_decorator(cmd(attribs=ProxyCreateCommandAttribs))
-    async def proxy_create(
-            self,
-            command: PlaygroundCommand) -> None:
-        # todo: add logging
-        if not await self.image_exists(command.data.image):
-            await self.pull_image(command.data.image)
-        _mappings = [
-            [m['mapping_from'], m['mapping_to']]
-            for m
-            in command.data.port_mappings]
-        mounts = {
-            "/etc/envoy": await self.volume_populate(
-                'proxy',
-                command.data.name,
-                'envoy',
-                {'envoy.yaml': base64.b64encode(
-                    command.data.configuration.encode('utf-8')).decode()}),
-            "/certs": await self.volume_populate(
-                'proxy',
-                command.data.name,
-                'certs',
-                command.data.certs.items()),
-            '/binary': await self.volume_populate(
-                'proxy',
-                command.data.name,
-                'binary',
-                command.data.binaries.items()),
-            '/logs': await self.volume_create(
-                'proxy', command.data.name, 'logs')}
-        container = await self.docker.containers.create_or_replace(
-            config=self._get_proxy_config(
-                command.data.image,
-                command.data.name,
-                mounts,
-                _mappings),
-            name="envoy__playground__proxy__%s" % command.data.name)
-        await container.start()
-
-    @method_decorator(cmd(attribs=ProxyDeleteAttribs))
-    async def proxy_delete(
-            self,
-            command: PlaygroundCommand) -> None:
-        # todo: use uuid
-        for container in await self.docker.containers.list():
-            if "envoy.playground.proxy" in container["Labels"]:
-                name_matches = (
-                    "/envoy__playground__proxy__%s" % command.data.name
-                    in container["Names"])
-                if name_matches:
-                    volumes = [
-                        v['Name']
-                        for v in container['Mounts']]
-                    await container.stop()
-                    await container.wait()
-                    await container.delete(v=True)
-                    if volumes:
-                        _volumes = await self.docker.volumes.list()
-                        for volume in _volumes['Volumes']:
-                            volume_name = volume['Name']
-                            if volume_name not in volumes:
-                                continue
-                            volume_delete = self.docker._query(
-                                    f"volumes/{volume_name}",
-                                    method="DELETE")
-                            async with volume_delete:
-                                pass
-
     async def pull_image(self, image_tag: str) -> None:
         if ":" not in image_tag:
             image_tag = f"{image_tag}:latest"
         await self.docker.images.pull(image_tag)
-
-    @method_decorator(cmd(attribs=ServiceCreateCommandAttribs))
-    async def service_create(
-            self,
-            command: PlaygroundCommand) -> None:
-        if not command.data.image:
-            # todo: add build logic
-            return
-        if not await self.image_exists(command.data.image):
-            await self.pull_image(command.data.image)
-        mounts = OrderedDict()
-        if command.data.configuration and command.data.config_path:
-            config = base64.b64encode(
-                command.data.configuration.encode('utf-8')).decode()
-            mounts[os.path.dirname(command.data.config_path)] = (
-                await self.volume_populate(
-                    'service',
-                    command.data.name,
-                    'config',
-                    {os.path.basename(command.data.config_path): config}))
-        _environment = [
-            "%s=%s" % (k, v)
-            for k, v
-            in command.data.env.items()]
-        container = await self.docker.containers.create_or_replace(
-            config=self._get_service_config(
-                command.data.service_type,
-                command.data.image,
-                command.data.name,
-                _environment,
-                mounts),
-            name="envoy__playground__service__%s" % command.data.name)
-        await container.start()
-
-    @method_decorator(cmd(attribs=ServiceDeleteAttribs))
-    async def service_delete(
-            self,
-            command: PlaygroundCommand) -> None:
-        for container in await self.docker.containers.list():
-            if "envoy.playground.service" in container["Labels"]:
-                name_matches = (
-                    "/envoy__playground__service__%s" % command.data.name
-                    in container["Names"])
-                if name_matches:
-                    volumes = [
-                        v['Name']
-                        for v in container['Mounts']]
-                    await container.stop()
-                    await container.wait()
-                    await container.delete(v=True, force=True)
-                    if volumes:
-                        _volumes = await self.docker.volumes.list()
-                        for volume in _volumes['Volumes']:
-                            volume_name = volume['Name']
-                            if volume_name not in volumes:
-                                continue
-                            volume_delete = self.docker._query(
-                                    f"volumes/{volume_name}",
-                                    method="DELETE")
-                            async with volume_delete:
-                                pass
 
     async def volume_create(
             self,
@@ -304,67 +150,6 @@ class PlaygroundDockerClient(object):
                 "envoy.playground.temp.mount": mount,
                 "envoy.playground.temp.target": target,
             }}
-
-    def _get_port_bindings(
-            self,
-            port_mappings: list) -> OrderedDict:
-        port_bindings: OrderedDict = OrderedDict()
-        for host, internal in port_mappings:
-            port_bindings[f"{internal}/tcp"] = port_bindings.get(
-                f"{internal}/tcp", [])
-            port_bindings[f"{internal}/tcp"].append(
-                {"HostPort": f"{host}"})
-        return port_bindings
-
-    def _get_proxy_config(
-            self,
-            image: str,
-            name: str,
-            mounts: dict,
-            port_mappings: list) -> dict:
-        # todo: handle udp etc
-        return {
-            'Image': image,
-            "AttachStdin": False,
-            "AttachStdout": False,
-            "AttachStderr": False,
-            "Tty": False,
-            "OpenStdin": False,
-            "Labels": {
-                "envoy.playground.proxy": name,
-            },
-            "HostConfig": {
-                "PortBindings": self._get_port_bindings(port_mappings),
-                "Binds": [
-                    '%s:%s' % (v.name, k)
-                    for k, v
-                    in mounts.items()]}}
-
-    def _get_service_config(
-            self,
-            service_type: str,
-            image: str,
-            name: str,
-            environment: list,
-            mounts: dict) -> dict:
-        labels = {
-            "envoy.playground.service": name,
-            "envoy.playground.service.type": service_type,
-        }
-        return {
-            'Image': image,
-            "AttachStdin": False,
-            "AttachStdout": False,
-            "AttachStderr": False,
-            "Tty": False,
-            "OpenStdin": False,
-            "Env": environment,
-            "Labels": labels,
-            "HostConfig": {
-                "Binds": [
-                    '%s:%s' % (v.name, k)
-                    for k, v
-                    in mounts.items()]}}
 
     async def _get_volume_config(
             self,
